:hammer: Fixing the code review request
